# Conflicts:
#	README.md
#	examples/mapSize8_60fps.gif

+++ food.cpp
@@ -2,7 +2,8 @@
 #include "thing.h"
 #include "food.h"
 
-Food::Food(int x, int y)
+Food::Food(double x, double y)
 	: Thing{ x,y,"food" }
 {
 }
+
